making alphabet component
